test(DatasourceEditor): remove describe() blocks and consolidate tests
Remove all describe() blocks from DatasourceEditor tests following the
codebase initiative to adopt flat test structure per Kent C. Dodds'
"avoid nesting when testing" best practices.

Changes:
- Merge DatasourceEditorRTL.test.tsx into DatasourceEditor.test.tsx
  - Eliminate misleading "RTL" naming (all files use React Testing Library)
  - Move 4 tests (metrics & columns) to main file
  - Remove 2 describe blocks
  - Result: 13 consolidated tests

- Refactor DatasourceEditorCurrency.test.tsx
  - Remove 1 describe block
  - Split mega-test into 3 focused tests
  - Each test now validates one specific behavior

- Refactor DashboardLinksExternal.test.tsx
  - Remove 1 describe block
  - Flatten 8 tests to root level

Results:
- 5 describe blocks removed across 3 files
- 1 redundant file deleted (RTL merged into main)
- 53 tests passing with 100% success rate
- Test structure follows modern best practices
- No conditional logic in tests
