+++ superset-frontend/src/components/Datasource/components/DatasourceEditor/components/DashboardLinksExternal/DashboardLinksExternal.test.tsx
@@ -46,88 +46,83 @@ const setupTest = (dashboards = mockDashboards) =>
     }),
   });
 
-// eslint-disable-next-line no-restricted-globals -- TODO: Migrate from describe blocks
-describe('DashboardLinksExternal', () => {
-  test('renders empty state when no dashboards provided', () => {
-    setupTest([]);
-    expect(screen.getByText('—')).toBeInTheDocument();
+test('renders empty state when no dashboards provided', () => {
+  setupTest([]);
+  expect(screen.getByText('—')).toBeInTheDocument();
+});
+
+test('renders empty state when dashboards is null/undefined', () => {
+  render(<DashboardLinksExternal dashboards={null as any} />, {
+    wrapper: createWrapper({
+      useRedux: true,
+      useRouter: true,
+    }),
   });
+  expect(screen.getByText('—')).toBeInTheDocument();
+});
 
-  test('renders empty state when dashboards is null/undefined', () => {
-    render(<DashboardLinksExternal dashboards={null as any} />, {
-      wrapper: createWrapper({
-        useRedux: true,
-        useRouter: true,
-      }),
-    });
-    expect(screen.getByText('—')).toBeInTheDocument();
+test('renders single dashboard link correctly', () => {
+  setupTest([mockDashboards[0]]);
+
+  const link = screen.getByText('Sales Dashboard');
+  expect(link).toBeInTheDocument();
+  expect(link.closest('a')).toHaveAttribute('href', '/superset/dashboard/1/');
+  expect(link.closest('a')).toHaveAttribute('target', '_blank');
+});
+
+test('renders multiple dashboard links with commas', () => {
+  setupTest();
+
+  expect(screen.getByText('Sales Dashboard')).toBeInTheDocument();
+  expect(screen.getByText(', Analytics Dashboard')).toBeInTheDocument();
+  expect(
+    screen.getByText(', Very Long Dashboard Name That Should Be Truncated'),
+  ).toBeInTheDocument();
+});
+
+test('all links open in new tabs', () => {
+  setupTest();
+
+  const links = screen.getAllByRole('link');
+  links.forEach(link => {
+    expect(link).toHaveAttribute('target', '_blank');
   });
+});
 
-  test('renders single dashboard link correctly', () => {
-    setupTest([mockDashboards[0]]);
+test('links have correct href attributes', () => {
+  setupTest();
 
-    const link = screen.getByText('Sales Dashboard');
-    expect(link).toBeInTheDocument();
-    expect(link.closest('a')).toHaveAttribute('href', '/superset/dashboard/1/');
-    expect(link.closest('a')).toHaveAttribute('target', '_blank');
-  });
+  const salesLink = screen.getByText('Sales Dashboard').closest('a');
+  const analyticsLink = screen.getByText(', Analytics Dashboard').closest('a');
+  const longNameLink = screen
+    .getByText(', Very Long Dashboard Name That Should Be Truncated')
+    .closest('a');
 
-  test('renders multiple dashboard links with commas', () => {
-    setupTest();
+  expect(salesLink).toHaveAttribute('href', '/superset/dashboard/1/');
+  expect(analyticsLink).toHaveAttribute('href', '/superset/dashboard/2/');
+  expect(longNameLink).toHaveAttribute('href', '/superset/dashboard/3/');
+});
 
-    expect(screen.getByText('Sales Dashboard')).toBeInTheDocument();
-    expect(screen.getByText(', Analytics Dashboard')).toBeInTheDocument();
-    expect(
-      screen.getByText(', Very Long Dashboard Name That Should Be Truncated'),
-    ).toBeInTheDocument();
-  });
+test('applies correct styling classes', () => {
+  setupTest();
 
-  test('all links open in new tabs', () => {
-    setupTest();
+  const truncatedSpan = document.querySelector('.truncated');
+  expect(truncatedSpan).toBeInTheDocument();
+  expect(truncatedSpan).toContainElement(screen.getAllByRole('link')[0]);
+});
 
-    const links = screen.getAllByRole('link');
-    links.forEach(link => {
-      expect(link).toHaveAttribute('target', '_blank');
-    });
-  });
+test('handles dashboard with empty title', () => {
+  const dashboardWithEmptyTitle = [
+    {
+      id: 1,
+      dashboard_title: '',
+      url: '/dashboard/1/',
+    },
+  ];
 
-  test('links have correct href attributes', () => {
-    setupTest();
+  setupTest(dashboardWithEmptyTitle);
 
-    const salesLink = screen.getByText('Sales Dashboard').closest('a');
-    const analyticsLink = screen
-      .getByText(', Analytics Dashboard')
-      .closest('a');
-    const longNameLink = screen
-      .getByText(', Very Long Dashboard Name That Should Be Truncated')
-      .closest('a');
-
-    expect(salesLink).toHaveAttribute('href', '/superset/dashboard/1/');
-    expect(analyticsLink).toHaveAttribute('href', '/superset/dashboard/2/');
-    expect(longNameLink).toHaveAttribute('href', '/superset/dashboard/3/');
-  });
-
-  test('applies correct styling classes', () => {
-    setupTest();
-
-    const truncatedSpan = document.querySelector('.truncated');
-    expect(truncatedSpan).toBeInTheDocument();
-    expect(truncatedSpan).toContainElement(screen.getAllByRole('link')[0]);
-  });
-
-  test('handles dashboard with empty title', () => {
-    const dashboardWithEmptyTitle = [
-      {
-        id: 1,
-        dashboard_title: '',
-        url: '/dashboard/1/',
-      },
-    ];
-
-    setupTest(dashboardWithEmptyTitle);
-
-    const link = screen.getByRole('link');
-    expect(link).toHaveTextContent('');
-    expect(link).toHaveAttribute('href', '/superset/dashboard/1/');
-  });
+  const link = screen.getByRole('link');
+  expect(link).toHaveTextContent('');
+  expect(link).toHaveAttribute('href', '/superset/dashboard/1/');
 });
